+++ numpy/_array_api_info.py
